Split API route handlers into focused domain modules
Extract 955 lines of route handlers from api.ts into 9 focused modules
under routes/. api.ts is now a slim server orchestrator (~417 lines,
down from 1,334).

New modules: metrics-router, tags-router, activities-router,
locations-router, settings-router, dashboard-router,
correlations-router, trends-router, and admin-router.

Also fixes route registration order for /metrics/bucketed and
/metrics/custom to come before /metrics/:metric, preventing
parameter capture of static paths.
