+++ apps/backend/src/api.ts
@@ -1,103 +1,10 @@
-/* eslint-disable max-lines -- TODO: refactor - split into route modules */
-import {
-  type ActivitiesQuery,
-  activitiesQuerySchema,
-  type ActivitiesResponse,
-  type ActivityImpactQuery,
-  activityImpactQuerySchema,
-  type ActivityImpactResponse,
-  type AddActivityBody,
-  addActivityBodySchema,
-  type AddActivityResponse,
-  type AddCustomMetricBody,
-  addCustomMetricBodySchema,
-  type AddMetricBody,
-  addMetricBodySchema,
-  type AddMetricResponse,
-  type AddNamedLocationBody,
-  addNamedLocationBodySchema,
-  type AddNamedLocationResponse,
-  type AddTagBody,
-  addTagBodySchema,
-  type AddTagResponse,
-  type AdminSettingsResponse,
-  type BaselineQuery,
-  baselineQuerySchema,
-  type BaselineResponse,
-  type BucketSize,
-  type CreateInvitationBody,
-  createInvitationBodySchema,
-  type CustomMetricResponse,
-  type CustomMetricsListResponse,
-  type DailySummaryQuery,
-  dailySummaryQuerySchema,
-  type DailySummaryResponse,
-  type DashboardResponse,
-  defaultDashboardConfig,
-  type DeleteActivityResponse,
-  type DeleteTagResponse,
-  type DetectedLocationsQuery,
-  detectedLocationsQuerySchema,
-  type DetectedLocationsResponse,
-  type EventProbabilityBody,
-  eventProbabilityBodySchema,
-  type EventProbabilityResponse,
-  type GenericCorrelationBody,
-  genericCorrelationBodySchema,
-  type GenericCorrelationResponse,
-  getExerciseTypeValue,
-  type GoalsProgressResponse,
-  type HrvActivitiesQuery,
-  hrvActivitiesQuerySchema,
-  type HrvActivitiesResponse,
-  type InvitationResponse,
-  isValidExerciseType,
-  type LocationsQuery,
-  locationsQuerySchema,
-  type LocationsResponse,
-  type LoginResponse,
-  type NamedLocationsResponse,
-  type PeriodSummaryQuery,
-  periodSummaryQuerySchema,
-  type PeriodSummaryResponse,
-  type ProductivityQuery,
-  productivityQuerySchema,
-  type ProductivityResponse,
-  type ProgrammaticTagsResponse,
-  type PromoteDetectedLocationBody,
-  promoteDetectedLocationBodySchema,
-  type QueryMetricsBucketedQuery,
-  queryMetricsBucketedQuerySchema,
-  type QueryMetricsBucketedResponse,
-  type QueryMetricsQuery,
-  queryMetricsQuerySchema,
-  type QueryMetricsResponse,
-  type ServerStatusResponse,
-  type SetTagMappingBody,
-  setTagMappingBodySchema,
-  type SetTagMappingResponse,
-  type SignupResponse,
-  type TagMappingsResponse,
-  type TagsQuery,
-  tagsQuerySchema,
-  type TagsResponse,
-  type TrendQuery,
-  trendQuerySchema,
-  type TrendResponse,
-  type UniqueTagsResponse,
-  type UpdateActivityBody,
-  updateActivityBodySchema,
-  type UpdateActivityResponse,
-  type UpdateAdminSettingsBody,
-  updateAdminSettingsBodySchema,
-  type UpdateDashboardInput,
-  updateDashboardInputSchema,
-  type UpdateNamedLocationBody,
-  updateNamedLocationBodySchema,
-  type UpdateSettingsInput,
-  updateSettingsInputSchema,
-  type UserSettingsResponse,
-} from '@aurboda/api-spec'
+/**
+ * Express server entry point.
+ *
+ * Route handlers are split into focused modules under routes/.
+ * This file handles: server setup, middleware, auth routes, and router mounting.
+ */
+import { type LoginResponse, type ServerStatusResponse, type SignupResponse } from '@aurboda/api-spec'
 import { json } from 'body-parser'
 import cors from 'cors'
 import express, { RequestHandler } from 'express'
@@ -106,10 +13,6 @@ import { createAuth } from './auth'
 import {
   getAllSyncStates,
   getDetectedLocationById,
-  getProgrammaticTags,
-  getDetectedLocations as getStoredDetectedLocations,
-  getUniqueTags,
-  getUserSettings,
   initializeSchema,
   insertLocation,
   insertPlace,
@@ -122,7 +25,6 @@ import {
   resetSyncState,
   schemaInitialized,
   updateDetectedLocation,
-  upsertUserSettings,
 } from './db'
 import { syncAllCalendars } from './ical-sync'
 import { createLastFmRouter } from './lastfm-router'
@@ -133,53 +35,23 @@ import { ouraClient } from './oura'
 import { syncAllOuraData } from './oura-sync'
 import { createOwnTracksRouter } from './owntracks'
 import { syncRescueTimeData } from './rescuetime-sync'
-import { isValidMetricOrCustom, type MetricType, validMetrics } from './schema'
+import { createActivitiesRouter } from './routes/activities-router'
+import { createAdminRouter } from './routes/admin-router'
+import { createCorrelationsRouter } from './routes/correlations-router'
+import { createDashboardRouter } from './routes/dashboard-router'
+import { createLocationsRouter } from './routes/locations-router'
+import { createMetricsRouter } from './routes/metrics-router'
+import { createSettingsRouter } from './routes/settings-router'
+import { createTagsRouter } from './routes/tags-router'
+import { createTrendsRouter } from './routes/trends-router'
 import { getCentralDb, initializeCentralDb } from './services/central-db'
-import {
-  getActivityImpact,
-  getBaseline,
-  getEventProbability,
-  getGenericCorrelation,
-  getHrvActivitiesCorrelation,
-} from './services/correlations'
 import { createDetectionTrigger, DetectionTrigger } from './services/detection-trigger'
 import { runDetectionForUser } from './services/detection-worker'
 import { createGeocodeQueue, GeocodeQueue } from './services/geocode-queue'
-import { getGoalsProgress } from './services/goals'
 import { createInvitationAuth } from './services/invitation'
-import {
-  deleteNamedLocation,
-  getDetectedLocations,
-  getNamedLocations,
-  insertNamedLocation,
-  updateNamedLocation,
-} from './services/locations'
-import {
-  addActivity,
-  addCustomMetric,
-  addMetric,
-  addTag,
-  deleteActivity,
-  deleteCustomMetric,
-  deleteTag,
-  getCustomMetrics,
-  updateActivity,
-} from './services/mutations'
-import {
-  getDailySummary,
-  getPeriodSummary,
-  queryActivities,
-  queryLocations,
-  queryMetrics,
-  queryMetricsBucketed,
-  queryProductivity,
-  queryTags,
-} from './services/queries'
-import { getSettings, getSettingsResponse, validateAndUpdateSettings } from './services/settings'
+import { getSettings } from './services/settings'
 import { createSyncProvider } from './services/sync-provider'
-import { getTrend } from './services/trends'
 import { createSyncRouter } from './sync-router'
-import { validateBody, validateQuery } from './validation'
 
 declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
@@ -190,7 +62,7 @@ declare global {
   }
 }
 
-// eslint-disable-next-line max-lines-per-function -- TODO: refactor - extract route groups
+// eslint-disable-next-line max-lines-per-function -- server setup and auth routes
 const main = async () => {
   const unauthorized = Object.assign(new Error('Unauthorized'), { status: 401 })
   const forbidden = Object.assign(new Error('Forbidden'), { status: 403 })
@@ -260,6 +132,10 @@ const main = async () => {
     next()
   }
 
+  // ==========================================================================
+  // Auth routes (stay here - tightly coupled to server setup)
+  // ==========================================================================
+
   httpd.get<Record<string, never>, ServerStatusResponse>('/status', async (_req, res) => {
     const signupMode = await centralDb.getSignupMode()
     res.json({
@@ -269,7 +145,7 @@ const main = async () => {
     })
   })
 
-  // eslint-disable-next-line complexity -- TODO: refactor
+  // eslint-disable-next-line complexity -- signup validation logic
   httpd.post<Record<string, never>, SignupResponse>('/signup', async (req, res, next) => {
     const signupMode = await centralDb.getSignupMode()
 
@@ -388,6 +264,10 @@ const main = async () => {
     res.end(JSON.stringify({ refresh, token: refresh }))
   })
 
+  // ==========================================================================
+  // External service routers (already extracted)
+  // ==========================================================================
+
   // Transform SyncState to ProviderSyncStatus format (undefined -> null)
   const transformSyncStates = async (user: string, provider: string) => {
     const states = await getAllSyncStates(user, provider)
@@ -400,7 +280,6 @@ const main = async () => {
     }))
   }
 
-  // Transform Oura sync results (Date -> ISO string)
   const transformOuraSyncResults = async (
     user: string,
     options: { fullResync?: boolean; startDate?: Date },
@@ -412,7 +291,6 @@ const main = async () => {
     }))
   }
 
-  // Transform RescueTime sync result (Date -> ISO string)
   const transformRescueTimeSyncResult = async (
     user: string,
     apiKey: string,
@@ -425,7 +303,6 @@ const main = async () => {
     }
   }
 
-  // Transform Last.fm sync result
   const transformLastFmSyncResult = async (
     user: string,
     apiKey: string,
@@ -435,7 +312,6 @@ const main = async () => {
     return await syncLastFmData(user, apiKey, username, options)
   }
 
-  // Sync router - handles /sync/* endpoints
   httpd.use(
     '/sync',
     createSyncRouter(
@@ -460,7 +336,6 @@ const main = async () => {
     ),
   )
 
-  // Last.fm router - handles /lastfm/* endpoints
   httpd.use('/lastfm', createLastFmRouter(authMiddleware))
 
   httpd.get('/auth/connectOura', oura.redirectToAuthorize)
@@ -484,7 +359,6 @@ const main = async () => {
     runDetectionForUser,
   })
 
-  // OwnTracks data endpoint (protected by Basic Auth using existing user credentials)
   httpd.use(
     '/ownTracks',
     createOwnTracksRouter({
@@ -496,813 +370,22 @@ const main = async () => {
   )
 
   // ==========================================================================
-  // REST API - Uses shared service layer with MCP
+  // REST API route modules
   // ==========================================================================
 
-  // GET /metrics/:metric - Query time series metrics
-  httpd.get<{ metric: string }, QueryMetricsResponse, unknown, QueryMetricsQuery>(
-    '/metrics/:metric',
-    authMiddleware,
-    validateQuery(queryMetricsQuerySchema),
-    async (req, res) => {
-      const { metric } = req.params
-      const { start, end } = req.query
-      const user = req.user!
-
-      const settings = await getUserSettings(user)
-      const customMetrics = settings?.customMetrics ?? []
-
-      if (!isValidMetricOrCustom(metric, customMetrics)) {
-        return res.status(400).json({
-          error: `Invalid metric "${metric}". Valid metrics are: ${validMetrics.join(', ')}`,
-          success: false,
-        })
-      }
-
-      const result = await queryMetrics(user, metric, new Date(start), new Date(end), customMetrics)
-      res.json({ ...result, success: true })
-    },
-  )
-
-  // Valid bucket sizes for bucketed metrics queries
-  const validBucketSizes = ['5m', '15m', '30m', '1h', '1d'] as const
-
-  // GET /metrics/bucketed - Query pre-aggregated metrics in time buckets
-  httpd.get<Record<string, never>, QueryMetricsBucketedResponse, unknown, QueryMetricsBucketedQuery>(
-    '/metrics/bucketed',
-    authMiddleware,
-    validateQuery(queryMetricsBucketedQuerySchema),
-    async (req, res) => {
-      const { start, end, bucket, metrics: metricsParam } = req.query
-      const user = req.user!
-
-      const settings = await getUserSettings(user)
-      const customMetrics = settings?.customMetrics ?? []
-
-      const metrics = metricsParam.split(',')
-      const invalidMetrics = metrics.filter((m) => !isValidMetricOrCustom(m, customMetrics))
-      if (invalidMetrics.length > 0) {
-        return res.status(400).json({
-          error: `Invalid metrics: ${invalidMetrics.join(', ')}. Valid metrics are: ${validMetrics.join(', ')}`,
-          success: false,
-        })
-      }
-
-      if (!validBucketSizes.includes(bucket)) {
-        return res.status(400).json({
-          error: `Invalid bucket size "${bucket}". Valid sizes are: ${validBucketSizes.join(', ')}`,
-          success: false,
-        })
-      }
-
-      const result = await queryMetricsBucketed(
-        user,
-        metrics as MetricType[],
-        new Date(start),
-        new Date(end),
-        bucket as BucketSize,
-      )
-      res.json({ ...result, success: true })
-    },
-  )
-
-  // GET /daily-summary - Get comprehensive summary for a day
-  httpd.get<Record<string, never>, DailySummaryResponse, unknown, DailySummaryQuery>(
-    '/daily-summary',
-    authMiddleware,
-    validateQuery(dailySummaryQuerySchema),
-    async (req, res) => {
-      const { date } = req.query
-      const user = req.user!
-
-      const summary = await getDailySummary(user, new Date(date), syncProvider)
-      res.json({ data: summary, success: true })
-    },
-  )
-
-  // GET /period-summary - Get aggregated stats for a period
-  httpd.get<Record<string, never>, PeriodSummaryResponse, unknown, PeriodSummaryQuery>(
-    '/period-summary',
-    authMiddleware,
-    validateQuery(periodSummaryQuerySchema),
-    async (req, res) => {
-      const { start, end, metrics: metricsParam } = req.query
-      const user = req.user!
-
-      const settings = await getUserSettings(user)
-      const customMetrics = settings?.customMetrics ?? []
-
-      const metrics = metricsParam.split(',')
-      const invalidMetrics = metrics.filter((m) => !isValidMetricOrCustom(m, customMetrics))
-      if (invalidMetrics.length > 0) {
-        return res.status(400).json({
-          error: `Invalid metrics: ${invalidMetrics.join(', ')}. Valid metrics are: ${validMetrics.join(', ')}`,
-          success: false,
-        })
-      }
-
-      const summary = await getPeriodSummary(user, metrics, new Date(start), new Date(end))
-      res.json({ ...summary, success: true })
-    },
-  )
-
-  // GET /tags - Query tags for a time range
-  httpd.get<Record<string, never>, TagsResponse, unknown, TagsQuery>(
-    '/tags',
-    authMiddleware,
-    validateQuery(tagsQuerySchema),
-    async (req, res) => {
-      const { start, end } = req.query
-      const user = req.user!
-
-      const tags = await queryTags(user, new Date(start), new Date(end), syncProvider)
-      res.json({ data: tags, success: true })
-    },
-  )
-
-  // POST /tags - Add a manual tag
-  httpd.post<Record<string, never>, AddTagResponse, AddTagBody>(
-    '/tags',
-    authMiddleware,
-    validateBody(addTagBodySchema),
-    async (req, res) => {
-      const { tag, start_time, end_time, merge_span } = req.body
-      const user = req.user!
-
-      const startDate = new Date(start_time)
-      const endDate = end_time ? new Date(end_time) : undefined
-
-      const result = await addTag(user, {
-        endTime: endDate,
-        mergeSpan: merge_span,
-        startTime: startDate,
-        tag,
-      })
-      res.json(result)
-    },
-  )
-
-  // DELETE /tags/:externalId - Delete a tag by external ID
-  httpd.delete<{ externalId: string }, DeleteTagResponse>(
-    '/tags/:externalId',
-    authMiddleware,
-    async (req, res) => {
-      const { externalId } = req.params
-      const user = req.user!
-
-      const result = await deleteTag(user, externalId)
-      res.json(result)
-    },
-  )
-
-  // GET /tags/unique - Get all unique tag names
-  httpd.get<Record<string, never>, UniqueTagsResponse>('/tags/unique', authMiddleware, async (req, res) => {
-    const user = req.user!
-    const tags = await getUniqueTags(user)
-    res.json({ data: tags, success: true })
-  })
-
-  // GET /tags/programmatic - Get all programmatic tags (UUIDs, tag_* prefixes) with their current mappings
-  httpd.get<Record<string, never>, ProgrammaticTagsResponse>(
-    '/tags/programmatic',
-    authMiddleware,
-    async (req, res) => {
-      const user = req.user!
-      const tags = await getProgrammaticTags(user)
-      const settings = await getUserSettings(user)
-      const mappings = settings?.tagMappings ?? {}
-
-      const data = tags.map((tag) => ({
-        count: tag.count,
-        currentName: mappings[tag.tagKey] ?? null,
-        latestTime: tag.latestTime.toISOString(),
-        tagKey: tag.tagKey,
-      }))
-
-      res.json({ data, success: true })
-    },
-  )
-
-  // POST /tags/mapping - Set a tag mapping
-  httpd.post<Record<string, never>, SetTagMappingResponse, SetTagMappingBody>(
-    '/tags/mapping',
-    authMiddleware,
-    validateBody(setTagMappingBodySchema),
-    async (req, res) => {
-      const { tagKey, name } = req.body
-      const user = req.user!
-
-      const settings = await getUserSettings(user)
-      const currentMappings = settings?.tagMappings ?? {}
-      const newMappings = { ...currentMappings, [tagKey]: name }
-
-      await upsertUserSettings(user, { tagMappings: newMappings })
-
-      res.json({ mapping: newMappings, success: true })
-    },
-  )
-
-  // GET /tags/mappings - Get all tag mappings
-  httpd.get<Record<string, never>, TagMappingsResponse>(
-    '/tags/mappings',
-    authMiddleware,
-    async (req, res) => {
-      const user = req.user!
-      const settings = await getUserSettings(user)
-      res.json({ mappings: settings?.tagMappings ?? {}, success: true })
-    },
-  )
-
-  // GET /activities - Query activities for a time range
-  httpd.get<Record<string, never>, ActivitiesResponse, unknown, ActivitiesQuery>(
-    '/activities',
-    authMiddleware,
-    validateQuery(activitiesQuerySchema),
-    async (req, res) => {
-      const { start, end, types: typesParam } = req.query
-      const user = req.user!
-
-      const types = (typesParam?.split(',') || ['sleep', 'exercise', 'meditation']) as (
-        | 'sleep'
-        | 'exercise'
-        | 'meditation'
-        | 'nap'
-      )[]
-
-      const activities = await queryActivities(user, types, new Date(start), new Date(end), syncProvider)
-      res.json({ data: activities, success: true })
-    },
-  )
-
-  // POST /activities - Add a manual activity
-  httpd.post<Record<string, never>, AddActivityResponse, AddActivityBody>(
-    '/activities',
-    authMiddleware,
-    validateBody(addActivityBodySchema),
-    async (req, res) => {
-      const { activity_type, start_time, end_time, title, notes, exercise_type } = req.body
-      const user = req.user!
-
-      const startDate = new Date(start_time)
-      const endDate = new Date(end_time)
-
-      // Validate and convert exercise_type name to value if provided
-      let data: Record<string, unknown> | undefined
-      if (exercise_type !== undefined) {
-        if (!isValidExerciseType(exercise_type)) {
-          return res.status(400).json({
-            error: `Invalid exercise_type "${exercise_type}"`,
-            success: false,
-          })
-        }
-        data = {
-          exerciseType: getExerciseTypeValue(exercise_type),
-          exerciseTypeName: exercise_type,
-        }
-      }
-
-      const result = await addActivity(user, {
-        activityType: activity_type,
-        data,
-        endTime: endDate,
-        notes,
-        startTime: startDate,
-        title,
-      })
-
-      if (!result.success) {
-        return res.status(400).json({ error: result.error, success: false })
-      }
-
-      res.json({
-        data: {
-          activityType: result.activityType!,
-          endTime: result.endTime!,
-          id: result.id!,
-          notes: result.notes,
-          startTime: result.startTime!,
-          title: result.title,
-        },
-        success: true,
-      })
-    },
-  )
-
-  // DELETE /activities/:id - Delete an activity by ID
-  httpd.delete<{ id: string }, DeleteActivityResponse>(
-    '/activities/:id',
-    authMiddleware,
-    async (req, res) => {
-      const { id } = req.params
-      const user = req.user!
-
-      const result = await deleteActivity(user, id)
-
-      if (!result.success) {
-        return res.status(404).json({ error: 'Activity not found', success: false })
-      }
-
-      res.json({ success: true })
-    },
-  )
-
-  // PATCH /activities/:id - Update an activity by ID
-  httpd.patch<{ id: string }, UpdateActivityResponse, UpdateActivityBody>(
-    '/activities/:id',
-    authMiddleware,
-    validateBody(updateActivityBodySchema),
-    async (req, res) => {
-      const { id } = req.params
-      const { start_time, end_time, title, notes } = req.body
-      const user = req.user!
-
-      const result = await updateActivity(user, id, {
-        endTime: end_time ? new Date(end_time) : undefined,
-        notes,
-        startTime: start_time ? new Date(start_time) : undefined,
-        title,
-      })
-
-      if (!result.success) {
-        const status = result.error === 'Activity not found' ? 404 : 400
-        return res.status(status).json({ error: result.error, success: false })
-      }
-
-      res.json({
-        data: {
-          activityType: result.activityType!,
-          endTime: result.endTime!,
-          id: result.id!,
-          notes: result.notes,
-          startTime: result.startTime!,
-          title: result.title,
-        },
-        success: true,
-      })
-    },
-  )
-
-  // GET /productivity - Query productivity data for a time range
-  httpd.get<Record<string, never>, ProductivityResponse, unknown, ProductivityQuery>(
-    '/productivity',
-    authMiddleware,
-    validateQuery(productivityQuerySchema),
-    async (req, res) => {
-      const { start, end } = req.query
-      const user = req.user!
-
-      const productivity = await queryProductivity(user, new Date(start), new Date(end), syncProvider)
-      res.json({ data: productivity, success: true })
-    },
-  )
-
-  // GET /locations - Query location data for a time range
-  httpd.get<Record<string, never>, LocationsResponse, unknown, LocationsQuery>(
-    '/locations',
-    authMiddleware,
-    validateQuery(locationsQuerySchema),
-    async (req, res) => {
-      const { start, end } = req.query
-      const user = req.user!
-
-      const places = await queryLocations(user, new Date(start), new Date(end))
-      res.json({ data: places, success: true })
-    },
-  )
+  httpd.use(createMetricsRouter(authMiddleware, syncProvider))
+  httpd.use('/tags', createTagsRouter(authMiddleware, syncProvider))
+  httpd.use(createActivitiesRouter(authMiddleware, syncProvider))
+  httpd.use('/locations', createLocationsRouter(authMiddleware))
+  httpd.use(createSettingsRouter(authMiddleware))
+  httpd.use('/dashboard', createDashboardRouter(authMiddleware))
+  httpd.use('/correlations', createCorrelationsRouter(authMiddleware, syncProvider))
+  httpd.use('/trends', createTrendsRouter(authMiddleware))
+  httpd.use('/admin', createAdminRouter(authMiddleware, adminMiddleware, centralDb, invitationAuth, webHost))
 
   // ==========================================================================
-  // Named Locations API
+  // Server startup
   // ==========================================================================
-
-  // GET /locations/named - List all named locations
-  httpd.get<Record<string, never>, NamedLocationsResponse>(
-    '/locations/named',
-    authMiddleware,
-    async (req, res) => {
-      const locations = await getNamedLocations(req.user!)
-      res.json({ data: locations, success: true })
-    },
-  )
-
-  // POST /locations/named - Create a named location
-  httpd.post<Record<string, never>, AddNamedLocationResponse, AddNamedLocationBody>(
-    '/locations/named',
-    authMiddleware,
-    validateBody(addNamedLocationBodySchema),
-    async (req, res) => {
-      const { name, lat, lon, radius } = req.body
-      const user = req.user!
-
-      const location = await insertNamedLocation(user, { lat, lon, name, radius })
-      res.json({ data: location, success: true })
-    },
-  )
-
-  // PATCH /locations/named/:id - Update a named location
-  httpd.patch<{ id: string }, AddNamedLocationResponse, UpdateNamedLocationBody>(
-    '/locations/named/:id',
-    authMiddleware,
-    validateBody(updateNamedLocationBodySchema),
-    async (req, res) => {
-      const { id } = req.params
-      const { name, lat, lon, radius } = req.body
-      const user = req.user!
-
-      // lat and lon must be updated together
-      if ((lat !== undefined) !== (lon !== undefined)) {
-        return res.status(400).json({ error: 'lat and lon must be updated together', success: false })
-      }
-
-      const location = await updateNamedLocation(user, id, { lat, lon, name, radius })
-      if (!location) {
-        return res.status(404).json({ error: 'Named location not found', success: false })
-      }
-      res.json({ data: location, success: true })
-    },
-  )
-
-  // DELETE /locations/named/:id - Delete a named location
-  httpd.delete<{ id: string }, DeleteTagResponse>(
-    '/locations/named/:id',
-    authMiddleware,
-    async (req, res) => {
-      const { id } = req.params
-      const deleted = await deleteNamedLocation(req.user!, id)
-      if (!deleted) {
-        return res.status(404).json({ error: 'Named location not found', success: false })
-      }
-      res.json({ success: true })
-    },
-  )
-
-  // GET /locations/detected - Get computed detected location clusters (on-demand analysis)
-  httpd.get<Record<string, never>, DetectedLocationsResponse, unknown, DetectedLocationsQuery>(
-    '/locations/detected',
-    authMiddleware,
-    validateQuery(detectedLocationsQuerySchema),
-    async (req, res) => {
-      const { start, end, min_duration } = req.query
-      const user = req.user!
-
-      const detected = await getDetectedLocations(user, {
-        end: new Date(end),
-        minDurationMinutes: min_duration ? parseInt(min_duration, 10) : undefined,
-        start: new Date(start),
-      })
-      res.json({ data: detected, success: true })
-    },
-  )
-
-  // GET /locations/detected/stored - Get stored detected locations with addresses
-  httpd.get<Record<string, never>, DetectedLocationsResponse>(
-    '/locations/detected/stored',
-    authMiddleware,
-    async (req, res) => {
-      const user = req.user!
-      const detected = await getStoredDetectedLocations(user)
-      // Transform Date objects to ISO strings for API response
-      const serialized = detected.map((d) => ({
-        ...d,
-        firstVisit: d.firstVisit.toISOString(),
-        lastVisit: d.lastVisit.toISOString(),
-      }))
-      res.json({ data: serialized, success: true })
-    },
-  )
-
-  // POST /locations/detected/promote - Promote detected location to named
-  httpd.post<Record<string, never>, AddNamedLocationResponse, PromoteDetectedLocationBody>(
-    '/locations/detected/promote',
-    authMiddleware,
-    validateBody(promoteDetectedLocationBodySchema),
-    async (req, res) => {
-      const { lat, lon, name, radius } = req.body
-      const user = req.user!
-
-      const location = await insertNamedLocation(user, { lat, lon, name, radius })
-      res.json({ data: location, success: true })
-    },
-  )
-
-  // GET /metrics/custom - List all custom metric types
-  httpd.get<Record<string, never>, CustomMetricsListResponse>(
-    '/metrics/custom',
-    authMiddleware,
-    async (req, res) => {
-      const user = req.user!
-      const metrics = await getCustomMetrics(user)
-      res.json({ data: metrics, success: true })
-    },
-  )
-
-  // POST /metrics/custom - Register a new custom metric type
-  httpd.post<Record<string, never>, CustomMetricResponse, AddCustomMetricBody>(
-    '/metrics/custom',
-    authMiddleware,
-    validateBody(addCustomMetricBodySchema),
-    async (req, res) => {
-      const user = req.user!
-      const result = await addCustomMetric(user, req.body)
-      if (!result.success) {
-        return res.status(400).json({ error: result.error, success: false })
-      }
-      res.json({ data: result.data, success: true })
-    },
-  )
-
-  // DELETE /metrics/custom/:name - Delete a custom metric type
-  httpd.delete<{ name: string }, CustomMetricResponse>(
-    '/metrics/custom/:name',
-    authMiddleware,
-    async (req, res) => {
-      const { name } = req.params
-      const user = req.user!
-      const result = await deleteCustomMetric(user, name)
-      if (!result.deleted) {
-        return res.status(404).json({ error: `Custom metric "${name}" not found`, success: false })
-      }
-      res.json({ success: true })
-    },
-  )
-
-  // POST /metrics - Add a manual metric measurement
-  httpd.post<Record<string, never>, AddMetricResponse, AddMetricBody>(
-    '/metrics',
-    authMiddleware,
-    validateBody(addMetricBodySchema),
-    async (req, res) => {
-      const { metric, value, time } = req.body
-      const user = req.user!
-
-      const measurementTime = time ? new Date(time) : new Date()
-
-      const result = await addMetric(user, { metric, time: measurementTime, value })
-      res.json(result)
-    },
-  )
-
-  // GET /goals/progress - Get progress toward all user goals
-  httpd.get<Record<string, never>, GoalsProgressResponse>(
-    '/goals/progress',
-    authMiddleware,
-    async (req, res) => {
-      const goals = await getGoalsProgress(req.user!)
-      res.json({ goals, success: true })
-    },
-  )
-
-  // GET /user/settings - Get user settings with effective HR zones
-  httpd.get<Record<string, never>, UserSettingsResponse>(
-    '/user/settings',
-    authMiddleware,
-    async (req, res) => {
-      const result = await getSettingsResponse(req.user!)
-      res.json(result)
-    },
-  )
-
-  // PATCH /user/settings - Update user settings
-  httpd.patch<Record<string, never>, UserSettingsResponse, UpdateSettingsInput>(
-    '/user/settings',
-    authMiddleware,
-    validateBody(updateSettingsInputSchema),
-    async (req, res) => {
-      const result = await validateAndUpdateSettings(req.user!, req.body)
-      if (!result.success) {
-        return res.status(400).json(result)
-      }
-      res.json(result)
-    },
-  )
-
-  // ==========================================================================
-  // Dashboard API
-  // ==========================================================================
-
-  // GET /dashboard - Get user's dashboard configuration
-  httpd.get<Record<string, never>, DashboardResponse>('/dashboard', authMiddleware, async (req, res) => {
-    const settings = await getSettings(req.user!)
-    const dashboard = settings.dashboard ?? defaultDashboardConfig
-    res.json({ dashboard, success: true })
-  })
-
-  // PUT /dashboard - Replace entire dashboard configuration
-  httpd.put<Record<string, never>, DashboardResponse, UpdateDashboardInput>(
-    '/dashboard',
-    authMiddleware,
-    validateBody(updateDashboardInputSchema),
-    async (req, res) => {
-      const user = req.user!
-      await upsertUserSettings(user, { dashboard: req.body })
-      res.json({ dashboard: req.body, success: true })
-    },
-  )
-
-  // POST /dashboard/reset - Reset dashboard to default configuration
-  httpd.post<Record<string, never>, DashboardResponse>(
-    '/dashboard/reset',
-    authMiddleware,
-    async (req, res) => {
-      const user = req.user!
-      await upsertUserSettings(user, { dashboard: undefined })
-      res.json({ dashboard: defaultDashboardConfig, success: true })
-    },
-  )
-
-  // ==========================================================================
-  // Correlations API
-  // ==========================================================================
-
-  // GET /correlations/baseline - Get HRV baseline statistics
-  httpd.get<Record<string, never>, BaselineResponse, unknown, BaselineQuery>(
-    '/correlations/baseline',
-    authMiddleware,
-    validateQuery(baselineQuerySchema),
-    async (req, res) => {
-      const { reference_date } = req.query
-      const user = req.user!
-
-      const referenceDate = reference_date ? new Date(reference_date) : undefined
-      const baseline = await getBaseline(user, referenceDate)
-      res.json({ data: baseline, success: true })
-    },
-  )
-
-  // GET /correlations/hrv-activities - Get HRV correlations with activities
-  httpd.get<Record<string, never>, HrvActivitiesResponse, unknown, HrvActivitiesQuery>(
-    '/correlations/hrv-activities',
-    authMiddleware,
-    validateQuery(hrvActivitiesQuerySchema),
-    async (req, res) => {
-      const { period_days } = req.query
-      const user = req.user!
-
-      const periodDays = period_days ? parseInt(period_days, 10) : 30
-      const correlations = await getHrvActivitiesCorrelation(user, periodDays, syncProvider)
-      res.json({ data: correlations, success: true })
-    },
-  )
-
-  // GET /correlations/activity-impact/:activity - Get activity impact on metrics
-  httpd.get<{ activity: string }, ActivityImpactResponse, unknown, ActivityImpactQuery>(
-    '/correlations/activity-impact/:activity',
-    authMiddleware,
-    validateQuery(activityImpactQuerySchema),
-    async (req, res) => {
-      const { activity } = req.params
-      const { activity_type, period_days, window_minutes } = req.query
-      const user = req.user!
-
-      const periodDays = period_days ? parseInt(period_days, 10) : 90
-      const windowMinutes = window_minutes ? parseInt(window_minutes, 10) : 30
-
-      const impact = await getActivityImpact(
-        user,
-        activity,
-        activity_type,
-        windowMinutes,
-        periodDays,
-        syncProvider,
-      )
-      res.json({ data: impact, success: true })
-    },
-  )
-
-  // POST /correlations/event-probability - Get event probability correlation
-  httpd.post<Record<string, never>, EventProbabilityResponse, EventProbabilityBody>(
-    '/correlations/event-probability',
-    authMiddleware,
-    validateBody(eventProbabilityBodySchema),
-    async (req, res) => {
-      const { trigger_type, trigger_value, outcome_pattern, lag_windows, period_days } = req.body
-      const user = req.user!
-
-      const probability = await getEventProbability(
-        user,
-        { type: trigger_type, value: trigger_value },
-        { pattern: outcome_pattern, type: 'tag' },
-        lag_windows ?? ['12h', '24h', '36h', '48h'],
-        period_days ?? 365,
-        syncProvider,
-      )
-      res.json({ data: probability, success: true })
-    },
-  )
-
-  // POST /correlations/generic - Generic correlation analysis
-  httpd.post<Record<string, never>, GenericCorrelationResponse, GenericCorrelationBody>(
-    '/correlations/generic',
-    authMiddleware,
-    validateBody(genericCorrelationBodySchema),
-    async (req, res) => {
-      const { triggers, outcome, lag_windows, period_days } = req.body
-      const user = req.user!
-
-      const result = await getGenericCorrelation(
-        user,
-        triggers,
-        outcome,
-        lag_windows ?? ['24h', '48h', '7d'],
-        period_days ?? 90,
-        syncProvider,
-      )
-      res.json({ data: result, success: true })
-    },
-  )
-
-  // ==========================================================================
-  // Trends API
-  // ==========================================================================
-
-  // GET /trends - Get time-weighted trend for tags or metrics
-  httpd.get<Record<string, never>, TrendResponse, unknown, TrendQuery>(
-    '/trends',
-    authMiddleware,
-    validateQuery(trendQuerySchema),
-    async (req, res) => {
-      const { aggregation, display_period, half_life_days, lookback_days, pattern, source_type } = req.query
-      const user = req.user!
-
-      const halfLifeDays = half_life_days ? parseInt(half_life_days, 10) : undefined
-      const lookbackDays = lookback_days ? parseInt(lookback_days, 10) : undefined
-
-      try {
-        const result = await getTrend(user, {
-          aggregation,
-          displayPeriod: display_period,
-          halfLifeDays,
-          lookbackDays,
-          pattern,
-          sourceType: source_type,
-        })
-        res.json({ data: result, success: true })
-      } catch (error) {
-        const message = error instanceof Error ? error.message : 'Unknown error'
-        res.status(400).json({ error: message, success: false })
-      }
-    },
-  )
-
-  // ==========================================================================
-  // Admin API
-  // ==========================================================================
-
-  // GET /admin/settings - Get admin settings
-  httpd.get<Record<string, never>, AdminSettingsResponse>(
-    '/admin/settings',
-    authMiddleware,
-    adminMiddleware,
-    async (_req, res) => {
-      const signupMode = await centralDb.getSignupMode()
-      const adminCount = await centralDb.getAdminCount()
-      res.json({
-        admin_count: adminCount,
-        signup_mode: signupMode,
-        success: true,
-      })
-    },
-  )
-
-  // PATCH /admin/settings - Update admin settings
-  httpd.patch<Record<string, never>, AdminSettingsResponse, UpdateAdminSettingsBody>(
-    '/admin/settings',
-    authMiddleware,
-    adminMiddleware,
-    validateBody(updateAdminSettingsBodySchema),
-    async (req, res) => {
-      const { signup_mode } = req.body
-      if (signup_mode) {
-        await centralDb.setSignupMode(signup_mode)
-      }
-      const currentMode = await centralDb.getSignupMode()
-      const adminCount = await centralDb.getAdminCount()
-      res.json({
-        admin_count: adminCount,
-        signup_mode: currentMode,
-        success: true,
-      })
-    },
-  )
-
-  // POST /admin/invitations - Create a new invitation
-  httpd.post<Record<string, never>, InvitationResponse, CreateInvitationBody>(
-    '/admin/invitations',
-    authMiddleware,
-    adminMiddleware,
-    validateBody(createInvitationBodySchema),
-    async (req, res) => {
-      const { expiryHours } = req.body
-      const token = invitationAuth.createInvitationToken(expiryHours)
-      const expiresAt = invitationAuth.getTokenExpiry(token)
-
-      res.json({
-        expiresAt: expiresAt!.toISOString(),
-        success: true,
-        token,
-        url: `${webHost}/signup?invite=${encodeURIComponent(token)}`,
-      })
-    },
-  )
 
   const port = Number(process.env.PORT ?? 80)
   const server = httpd.listen(port, () => {
