+++ apps/backend/src/routes/dashboard-router.ts
@@ -1,0 +1,47 @@
+/**
+ * Dashboard route group.
+ *
+ * Handles: /dashboard/*
+ */
+import {
+  type DashboardResponse,
+  defaultDashboardConfig,
+  type UpdateDashboardInput,
+  updateDashboardInputSchema,
+} from '@aurboda/api-spec'
+import { RequestHandler, Router } from 'express'
+import { upsertUserSettings } from '../db'
+import { getSettings } from '../services/settings'
+import { validateBody } from '../validation'
+
+export const createDashboardRouter = (authMiddleware: RequestHandler): Router => {
+  const router = Router()
+
+  // GET /dashboard - Get user's dashboard configuration
+  router.get<Record<string, never>, DashboardResponse>('/', authMiddleware, async (req, res) => {
+    const settings = await getSettings(req.user!)
+    const dashboard = settings.dashboard ?? defaultDashboardConfig
+    res.json({ dashboard, success: true })
+  })
+
+  // PUT /dashboard - Replace entire dashboard configuration
+  router.put<Record<string, never>, DashboardResponse, UpdateDashboardInput>(
+    '/',
+    authMiddleware,
+    validateBody(updateDashboardInputSchema),
+    async (req, res) => {
+      const user = req.user!
+      await upsertUserSettings(user, { dashboard: req.body })
+      res.json({ dashboard: req.body, success: true })
+    },
+  )
+
+  // POST /dashboard/reset - Reset dashboard to default configuration
+  router.post<Record<string, never>, DashboardResponse>('/reset', authMiddleware, async (req, res) => {
+    const user = req.user!
+    await upsertUserSettings(user, { dashboard: undefined })
+    res.json({ dashboard: defaultDashboardConfig, success: true })
+  })
+
+  return router
+}
